chore: deleted apps removed from list
